LRAC-15338 Remove old geomap components that uses clay-charts-shared (deprecated) and move utils to another folder
