+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/css/_account_overview.scss
@@ -23,9 +23,5 @@
 		.score {
 			width: 15%;
 		}
-
-		.trendline {
-			width: auto;
-		}
 	}
 }
