--- modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/clay-charts-react/Predictive.tsx
@@ -1,85 +1,0 @@
-import BillboardChart from 'react-billboardjs';
-import React from 'react';
-
-/**
- * Predictive Chart component.
- * @extends React.Component
- * @param {Object} props
- * @return {ReactElement}
- */
-
-interface IPredictiveChartProps {
-	data: Object;
-}
-
-type State = {
-	columns: string[] | number[];
-	regions: Object[];
-};
-
-function isDefAndNotNull(value) {
-	return value !== 'undefined' && value !== null;
-}
-
-export default class PredictiveChart extends React.Component<
-	IPredictiveChartProps,
-	State
-> {
-	/** @inheritdoc */
-	constructor(props) {
-		super(props);
-
-		let columns = props.data.columns;
-
-		if (isDefAndNotNull(columns)) {
-			columns = columns.map(dataSeries => {
-				if (dataSeries[0] !== 'x') {
-					dataSeries = dataSeries.map(element =>
-						typeof element === 'number'
-							? {
-									high: element,
-									low: element,
-									mid: element
-							  }
-							: element
-					);
-				}
-
-				return dataSeries;
-			});
-		}
-
-		const predictionDate = props.predictionDate;
-		const regions = [];
-
-		if (isDefAndNotNull(predictionDate)) {
-			regions.push({
-				axis: 'x',
-				start: predictionDate
-			});
-		}
-
-		this.state = {
-			...this.state,
-			columns,
-			regions
-		};
-	}
-
-	/** @inheritdoc */
-	render() {
-		const {columns, regions} = this.state;
-		const {data, ...otherProps} = this.props;
-
-		return (
-			<BillboardChart
-				data={{
-					...data,
-					columns
-				}}
-				regions={regions}
-				{...otherProps}
-			/>
-		);
-	}
-}
