+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/contacts/components/ActivitiesChartDeprecated.tsx
@@ -19,7 +19,7 @@ import {
 	XAxis,
 	YAxis
 } from 'recharts';
-import {CHART_COLOR_NAMES} from 'shared/components/Chart';
+import {CHART_COLOR_NAMES} from 'shared/util/charts';
 import {createDateKeysIMap} from 'shared/util/intervals';
 import {
 	formatXAxisDate,
