+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/shared/components/metric-card/__tests__/util.tsx
@@ -1,5 +1,5 @@
 import * as data from 'test/data';
-import {CHART_COLOR_NAMES} from 'shared/components/Chart';
+import {CHART_COLOR_NAMES} from 'shared/util/charts';
 import {CompositeMetric} from '../metrics';
 import {
 	convertHistogramKeysToDate,
