+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/shared/components/metric-card/util.tsx
@@ -1,7 +1,7 @@
 import MetricValue from './MetricValue';
 import React, {Fragment} from 'react';
 import Trend from 'shared/components/Trend';
-import {CHART_COLOR_NAMES} from 'shared/components/Chart';
+import {CHART_COLOR_NAMES} from 'shared/util/charts';
 import {get, last} from 'lodash';
 import {
 	getAxisFormatter,
