+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/sites/components/cohort-analysis/utils.tsx
@@ -1,5 +1,5 @@
 import moment from 'moment';
-import {CHART_COLOR_NAMES} from 'shared/components/Chart';
+import {CHART_COLOR_NAMES} from 'shared/util/charts';
 import {sub} from 'shared/util/lang';
 
 const {
